refactor(test): extract shared findElementByType util; trim WHAT comments (PR #527 review)
