+++ src/app/[symbol]/__tests__/page.test.ts
@@ -68,7 +68,7 @@ import { generateMetadata, default as SymbolPage } from '@/app/[symbol]/page';
 import { getAssetInfoCached } from '@/entities/ticker';
 import { peekAnalysisCache } from '@y0ngha/siglens-core';
 import { SymbolPageClient } from '@/widgets/symbol-page/SymbolPageClient';
-import { isValidElement, type ReactElement, type ReactNode } from 'react';
+import { findElementByType } from '@/shared/test-utils/findElementByType';
 import type { MockedFunction } from 'vitest';
 
 const mockGetAssetInfoCached = getAssetInfoCached as MockedFunction<
@@ -77,29 +77,6 @@ const mockGetAssetInfoCached = getAssetInfoCached as MockedFunction<
 const mockPeekAnalysisCache = peekAnalysisCache as MockedFunction<
     typeof peekAnalysisCache
 >;
-
-/**
- * 렌더 없이 RSC가 반환한 element 트리를 재귀 순회해 주어진 컴포넌트 타입의
- * 첫 element를 찾는다. async 서버 컴포넌트(page.tsx)는 @testing-library/react로
- * 직접 렌더할 수 없으므로(Promise<JSX.Element> 반환), props 주입 검증은 트리
- * 탐색으로 수행한다.
- */
-function findElementByType(
-    node: ReactNode,
-    type: unknown
-): ReactElement | null {
-    if (Array.isArray(node)) {
-        for (const child of node) {
-            const found = findElementByType(child, type);
-            if (found !== null) return found;
-        }
-        return null;
-    }
-    if (!isValidElement(node)) return null;
-    if (node.type === type) return node;
-    const childProps = node.props as { children?: ReactNode };
-    return findElementByType(childProps.children, type);
-}
 
 interface ClientSeedProps {
     initialAnalysis: unknown;
